logic name refactorings

+++ target-wasm/build.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-rm -rf ./pkg
+[ -e ./pkg ] && rm -rf ./pkg
 mkdir ./pkg
 
 set -e #fail on first error
@@ -8,6 +8,6 @@ wasm-pack build --release --target web
 cp ./pkg/target_wasm.js ./src
 cp ./pkg/target_wasm_bg.wasm ./src
 
-rm archive.zip
+[ -e archive.zip ] && rm archive.zip
 zip -r -j archive.zip ./src
 zip -d archive.zip lib.rs
